rename groups in profile
