+++ app/javascript/components/Profile/JoinedGroups.tsx
